+++ src/ui/UpgradePanel.js
@@ -136,6 +136,9 @@ export default class UpgradePanel {
     if (this.scene.prestigePanel?._isOpen) {
       this.scene.prestigePanel._close();
     }
+    if (this.scene.settingsPanel?._isOpen) {
+      this.scene.settingsPanel._close();
+    }
     this._isOpen = true;
     this._showModal();
     this._refresh();